requirements.txt
click==8.1.7
prettytable==3.10.0
python3-discogs-client==2.7
requests==2.31.0